# Loop over `diagind` in `diag` for banded matrices